requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-toolbelt==0.0.1
gunicorn==19.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
static3==0.7.0
whitenoise==4.1.2
